requirements/base.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    make upgrade
#
certifi==2020.6.20        # via requests
cffi==1.14.2              # via cryptography
chardet==3.0.4            # via requests
coreapi==2.3.3            # via django-rest-swagger, openapi-codec
coreschema==0.0.4         # via coreapi
cryptography==3.0         # via pyjwt, social-auth-core
defusedxml==0.6.0         # via python3-openid, social-auth-core
django-cors-headers==3.5.0  # via -r requirements/base.in
django-crum==0.7.7        # via edx-rbac
django-extensions==3.0.5  # via -r requirements/base.in
django-model-utils==4.0.0  # via edx-rbac
django-rest-swagger==2.2.0  # via -r requirements/base.in
django-waffle==1.0.0      # via -r requirements/base.in, edx-django-utils, edx-drf-extensions
django==2.2.15            # via -c requirements/constraints.txt, -r requirements/base.in, django-cors-headers, django-crum, django-model-utils, djangorestframework, drf-jwt, edx-auth-backends, edx-django-release-util, edx-django-utils, edx-drf-extensions, edx-rbac, rest-condition
djangorestframework==3.11.1  # via -r requirements/base.in, django-rest-swagger, drf-jwt, edx-drf-extensions, rest-condition
drf-jwt==1.17.0           # via edx-drf-extensions
edx-auth-backends==3.1.0  # via -r requirements/base.in
edx-django-release-util==0.4.4  # via -r requirements/base.in
edx-django-utils==3.7.3   # via -r requirements/base.in, edx-drf-extensions
edx-drf-extensions==6.1.1  # via -r requirements/base.in, edx-rbac
edx-opaque-keys==2.1.1    # via edx-drf-extensions
edx-rbac==1.3.2           # via -r requirements/base.in
edx-rest-api-client==1.9.2  # via -r requirements/base.in
future==0.18.2            # via pyjwkest
idna==2.10                # via requests
importlib-metadata==1.7.0  # via stevedore
itypes==1.2.0             # via coreapi
jinja2==2.11.2            # via coreschema
markupsafe==1.1.1         # via jinja2
mysqlclient==2.0.1        # via -r requirements/base.in
newrelic==5.16.2.147      # via edx-django-utils
oauthlib==3.1.0           # via requests-oauthlib, social-auth-core
openapi-codec==1.3.2      # via django-rest-swagger
pbr==5.4.5                # via stevedore
psutil==5.7.2             # via edx-django-utils
pycparser==2.20           # via cffi
pycryptodomex==3.9.8      # via pyjwkest
pyjwkest==1.4.2           # via edx-drf-extensions
pyjwt[crypto]==1.7.1      # via drf-jwt, edx-auth-backends, edx-rest-api-client, social-auth-core
pymongo==3.11.0           # via edx-opaque-keys
python-dateutil==2.8.1    # via edx-drf-extensions
python3-openid==3.2.0     # via social-auth-core
pytz==2020.1              # via -r requirements/base.in, django
pyyaml==5.3.1             # via edx-django-release-util
requests-oauthlib==1.3.0  # via social-auth-core
requests==2.24.0          # via coreapi, edx-drf-extensions, edx-rest-api-client, pyjwkest, requests-oauthlib, slumber, social-auth-core
rest-condition==1.0.3     # via edx-drf-extensions
rules==2.2                # via -r requirements/base.in
semantic-version==2.8.5   # via edx-drf-extensions
simplejson==3.17.2        # via django-rest-swagger
six==1.15.0               # via cryptography, edx-auth-backends, edx-django-release-util, edx-drf-extensions, edx-opaque-keys, edx-rbac, pyjwkest, python-dateutil, social-auth-app-django, social-auth-core
slumber==0.7.1            # via edx-rest-api-client
social-auth-app-django==4.0.0  # via edx-auth-backends
social-auth-core==3.3.3   # via edx-auth-backends, social-auth-app-django
sqlparse==0.3.1           # via django
stevedore==3.2.0          # via edx-django-utils, edx-opaque-keys
uritemplate==3.0.1        # via coreapi
urllib3==1.25.10          # via requests
zipp==3.1.0               # via importlib-metadata
